+++ example/lib/docs_ui/url_strategy_stub.dart
@@ -1,0 +1,5 @@
+/// The non-web half of the conditional import in `url_strategy.dart`.
+library;
+
+/// Does nothing: there is no URL to have a strategy about off the web.
+void useDocsUrlStrategy() {}